+++ dicom_validator/spec_reader/enum_parser.py
@@ -5,12 +5,14 @@ from typing import Optional
 
 from pydicom.valuerep import VR, INT_VR, STR_VR
 
+from dicom_validator.spec_reader.condition_parser import ConditionParser
+
 try:
     import lxml.etree as ElementTree
 except ImportError:
     import xml.etree.ElementTree as ElementTree
 
-from dicom_validator.spec_reader.condition import ValuesType
+from dicom_validator.spec_reader.condition import ValuesType, ConditionType, Condition
 
 OptionalElement = Optional[ElementTree.Element]
 
@@ -20,9 +22,18 @@ class EnumParser:
 
     docbook_ns = "{http://docbook.org/ns/docbook}"
     enum_with_value_regex = re.compile(r"Enumerated Values for Value (\d):")
+    enum_values_regex = re.compile(
+        r"[Ee]numerated [Vv]alues( (for|of) [A-Za-z0-9 ]+"
+        r"\([0-9A-Fa-f]{4},[0-9A-Fa-f]{4}\))?( (if|when) ([^:]*))?:?$"
+    )
 
-    def __init__(self, find_section: Callable[[str], OptionalElement]) -> None:
+    def __init__(
+        self,
+        find_section: Callable[[str], OptionalElement],
+        condition_parser: ConditionParser,
+    ) -> None:
         self._find_section = find_section
+        self._condition_parser = condition_parser
         self._enum_cache: dict[str, dict] = {}
 
     def parse(self, node: ElementTree.Element, vr: VR) -> list[dict]:
@@ -77,24 +88,31 @@ class EnumParser:
         title = var_list.find(self.docbook_ns + "title")
         if title is None:
             return {}
-        value_match = self.enum_with_value_regex.match(title.text)
-        if value_match:
-            index = int(value_match.group(1))
-        else:
-            index = 0
-        if index == 0 and title.text.lower() != "enumerated values:":
-            return {}
-        # TODO: handle cases with conditions
+        index = 0
+        match = self.enum_values_regex.match(title.text)
+        if not match:
+            value_match = self.enum_with_value_regex.match(title.text)
+            if value_match:
+                index = int(value_match.group(1))
+            else:
+                return {}
+        condition = None
+        if match and (cond := match.group(5)) is not None:
+            condition = self._condition_parser.parse("Required if " + cond)
+            if condition.type == ConditionType.UserDefined:
+                condition = None
         terms = []
         for item in var_list.findall(self.docbook_ns + "varlistentry"):
             term = item.find(self.docbook_ns + "term")
             if term is not None:
                 terms.append(term.text)
-        result: dict[str, list[str] | int] = {}
+        result: dict[str, list[str] | int | Condition] = {}
         if terms:
             result["val"] = terms
             if index > 0:
                 result["index"] = index
+            if condition is not None:
+                result["cond"] = condition
 
         return result